# Generates a go_plugin_build_manifest file when building a plugin