--- a/build/common.go
+++ b/build/common.go
@@ -23,6 +23,8 @@ import (
 	"github.com/urfave/cli"
 )
 
+var defaultOutputBinaryPath = "dist"
+
 // Callbacks give you a way to run custom behavior when things happen
 var beforeBuild = func(cfg Config) (Config, error) {
 	return cfg, nil
@@ -81,7 +83,7 @@ func buildBackend(cfg Config) error {
 
 	outputPath := cfg.OutputBinaryPath
 	if outputPath == "" {
-		outputPath = "dist"
+		outputPath = defaultOutputBinaryPath
 	}
 	args := []string{
 		"build", "-o", filepath.Join(outputPath, exeName),
@@ -168,6 +170,30 @@ func (Build) DarwinARM64() error {
 	return buildBackend(newBuildConfig("darwin", "arm64"))
 }
 
+// GenerateManifestFile generates a manifest file for plugin submissions
+func (Build) GenerateManifestFile() error {
+	config := Config{}
+	config, err := beforeBuild(config)
+	if err != nil {
+		return err
+	}
+	outputPath := config.OutputBinaryPath
+	if outputPath == "" {
+		outputPath = defaultOutputBinaryPath
+	}
+	manifestContent := utils.GenerateManifest()
+	manifestFile := filepath.Join(outputPath, "go_plugin_build_manifest")
+	err = os.MkdirAll(outputPath, 0755)
+	if err != nil {
+		return err
+	}
+	err = os.WriteFile(manifestFile, []byte(manifestContent), 0600)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
 // Debug builds the debug version for the current platform
 func (Build) Debug() error {
 	cfg := newBuildConfig(runtime.GOOS, runtime.GOARCH)
@@ -192,7 +218,7 @@ func (Build) Backend() error {
 // BuildAll builds production executables for all supported platforms.
 func BuildAll() { //revive:disable-line
 	b := Build{}
-	mg.Deps(b.Linux, b.Windows, b.Darwin, b.DarwinARM64, b.LinuxARM64, b.LinuxARM)
+	mg.Deps(b.Linux, b.Windows, b.Darwin, b.DarwinARM64, b.LinuxARM64, b.LinuxARM, b.GenerateManifestFile)
 }
 
 //go:embed tmpl/*