**What this PR does / why we need it**:

We (grafana) want to check the integrity of plugins binary files at submission time. 

We can use this manifest file to compare it against the source code we are provided and make sure no mistakes were made at submission.

Note that it is clear this manifest can be tampered with and this is not meant to be a fool-proof way of replicating builds. This is not a strict security feature against an intentional attack. This is oriented to prevent plugins submissions to mistakenly submit the wrong source code.

**Special notes for your reviewer**:

## Performance

Using [hyperfine](https://github.com/sharkdp/hyperfine) to run benchmarks against a [known plugin](https://github.com/alexanderzobnin/grafana-zabbix/) go code. 50 runs with 5 warm up runs (to fill the go build cache) and removing all dist files between runs:

### Zabbix datasource (around 4200 lines of code) 

#### With manifest generation

```
❯ hyperfine --runs 50 'mage buildAll' --warmup 5 --prepare 'rm -rf dist'
Benchmark 1: mage buildAll
  Time (mean ± σ):     945.4 ms ±  45.8 ms    [User: 5086.5 ms, System: 1636.5 ms]
  Range (min … max):   844.4 ms … 1049.5 ms    50 runs
```

#### Without manifest generation

```
❯ hyperfine --runs 50 'mage buildAll' --warmup 5 --prepare 'rm -rf dist'
Benchmark 1: mage buildAll
  Time (mean ± σ):     942.6 ms ±  49.5 ms    [User: 4885.4 ms, System: 1501.2 ms]
  Range (min … max):   843.3 ms … 1088.3 ms    50 runs
```

The difference in compile times speed were around 3ms and given the calculated margin of error this can be deemed as negligible.

### testing datadogs (larger code base ~6800 lines of code)
#### With manifest code
```
❯ hyperfine --runs 50 'mage buildAll' --prepare 'rm -rf dist' --warmup 3
Benchmark 1: mage buildAll
  Time (mean ± σ):      1.291 s ±  0.058 s    [User: 6.978 s, System: 1.789 s]
  Range (min … max):    1.191 s …  1.427 s    50 runs
```

#### Without manifest code
```
❯ hyperfine --runs 50 'mage buildAll' --prepare 'rm -rf dist' --warmup 3
Benchmark 1: mage buildAll
  Time (mean ± σ):      1.264 s ±  0.052 s    [User: 6.768 s, System: 1.690 s]
  Range (min … max):    1.185 s …  1.420 s    50 runs
```

### Using system commands instead of Walking and crypto.

A similar and much faster result can be achieved by using `find` and `sha1sum` commands but these are not compatible with windows development.