@academo 2023-02-08
> how do you assure that the manifest file haven't been tampered with after the build, but before publish?

We don't. This is not a strict security feature against an intentional attack. This is oriented to prevent plugins submissions to mistakenly submit the wrong source code.